web/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
pytz==2019.1
sqlparse==0.3.0
gunicorn==19.9.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
requests==2.22.0
beautifulsoup4==4.8.0